Repackage, rename: Dto, Mapper
