+++ src/main/java/com/app/car/rental/backend/service/mapper/avis/AvisApiReservationPostResponseMapper.java
@@ -1,0 +1,4 @@
+package com.app.car.rental.backend.service.mapper.avis;
+
+public class AvisApiReservationPostResponseMapper {
+}
