+++ src/main/java/com/app/car/rental/backend/web/model/request/CarChooseRequestDto.java
@@ -1,0 +1,14 @@
+package com.app.car.rental.backend.web.model.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CarChooseRequestDto {
+    private Long id;
+}
